@@ -2,7 +2,7 @@ package com.turki.alwaysonlibrarycontainer;
 
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import java.util.Calendar;
 
